requirements.txt

torch
transformers
diffusers
accelerate
safetensors
openai
streamlit
Pillow
